Yaw control bugs fixed. Improved debugging.
So I was flying my quadcopter. It was just hovering around minding its
own business, when it suddenly decided to ignore my transmitter and do a
front flip. Well, long story short, a propeller broke and the mount got
bent. And an even longer story even shorter, it was a hardware issue.

*sigh*