Strange motor behavior fixed.  PID stability improved.
An issue existed in the last commit where the ESCs would only output
minimal or maximal values. While I do not understand completely what the
underlying issue was, I have figured out that adding a delay between the
MPU6050 initialization and the MPU6050 update fixed the problem. My
guess is that the MPU6050 needed time to process the initialization
signals sent from the Arduino, so the system got messed up when the
Arduino was attempting to read from a pre-initialized MPU6050. Anyway,
keep an eye on it.

Also, I have fixed an issue in which the PID algorithms updated
inconsistently. In some iterations, they would compute, and in others
they would not. This is because the PID algorithms are designed to work
in a closed-loop where they are constantly checking to see if a certain
amount of time has passed, at which point they perform their
calculations. Since my loop is precisely timed, the PID algorithms would
often check 1 millisecond before a check-up was valid, causing it to
wait an entire cycle before it could update again. This caused variation
between the routines of the yaw, pitch, and roll PID algorithms,
creating instability. To resolve the issue, I have forced the loop to
wait until the PID algorithms are computed (usually 0 - 2 milliseconds)
before continuing to the next iteration.

Hopefully this new system works. I will perform some flight tests soon,
and if they are successful, I will merge this branch with the master.